@@ -1,8 +1,4 @@
-﻿using Bannerlord.ButterLib.Common.Helpers;
-
-using HarmonyLib;
-
-using Newtonsoft.Json;
+﻿using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
 
 using System;
@@ -14,36 +10,12 @@ using TaleWorlds.SaveSystem;
 
 namespace Bannerlord.ButterLib.SaveSystem
 {
-    internal class TaleWorldsContractResolver : DefaultContractResolver
+    public class TaleWorldsContractResolver : TaleWorldsContractResolverBase
     {
-        private delegate bool IsContainerDelegate(Type type);
-        private static readonly IsContainerDelegate? _isContainerDelegate =
-            AccessTools2.GetDelegate<IsContainerDelegate>(
-                AccessTools.Method(
-                    typeof(SaveableRootClassAttribute).Assembly.GetType("TaleWorlds.SaveSystem.TypeExtensions"), "IsContainer", new [] { typeof(Type) }));
-
-        private static bool IsContainerFallback(Type type)
-        {
-            if (type.IsGenericType && !type.IsGenericTypeDefinition)
-            {
-                var genericTypeDefinition = type.GetGenericTypeDefinition();
-                if (genericTypeDefinition == typeof(Dictionary<,>))
-                    return true;
-                if (genericTypeDefinition == typeof(List<>))
-                    return true;
-                if (genericTypeDefinition == typeof(Queue<>))
-                    return true;
-            }
-            else if (type.IsArray)
-                return true;
-
-            return false;
-        }
-
         protected override IList<JsonProperty> CreateProperties(Type type, MemberSerialization memberSerialization)
         {
             // Container are an exception.
-            if ((_isContainerDelegate != null && _isContainerDelegate(type)) || (_isContainerDelegate == null && IsContainerFallback(type)))
+            if ((_isContainerDelegate is not null && _isContainerDelegate(type)) || (_isContainerDelegate is null && IsContainerFallback(type)))
                 return base.CreateProperties(type, memberSerialization);
 
             // SaveableRootClassAttribute is not needed
@@ -56,11 +28,9 @@ namespace Bannerlord.ButterLib.SaveSystem
             }
 
             var fields = type.GetFields(BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic)
-                .Where(x => x.GetCustomAttributes(true).Any(a =>
-                    a.GetType() == typeof(SaveableFieldAttribute) || a.GetType() == typeof(SaveablePropertyAttribute)));
+                .Where(x => x.GetCustomAttribute<SaveableFieldAttribute>(true) is not null);
             var properties = type.GetProperties(BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic)
-                .Where(x => x.GetCustomAttributes(true).Any(a =>
-                    a.GetType() == typeof(SaveableFieldAttribute) || a.GetType() == typeof(SaveablePropertyAttribute)));
+                .Where(x => x.GetCustomAttribute<SaveablePropertyAttribute>(true) is not null);
 
             return fields.Select(p => new JsonProperty
             {
