@@ -2,6 +2,7 @@
 
 using Microsoft.Extensions.DependencyInjection;
 
+using System;
 using System.Diagnostics.CodeAnalysis;
 
 using TaleWorlds.ObjectSystem;
@@ -9,18 +10,45 @@ using TaleWorlds.ObjectSystem;
 namespace Bannerlord.ButterLib.ObjectSystem.Extensions
 {
     /// <summary>
-    /// <see cref="MBObjectBase"/> extension methods for dynamic variable / flag storage
+    /// <see cref="MBObjectBase"/> extension methods for Object Extension Data (dynamic variable / flag storage)
     /// </summary>
-    internal static class MBObjectBaseExtensions
+    public static class MBObjectBaseExtensions
     {
-        private static IMBObjectVariableStorage? _instance;
+        private static IMBObjectExtensionDataStore? _instance;
 
-        private static IMBObjectVariableStorage? Instance =>
-            _instance ??= ButterLibSubModule.Instance?.GetServiceProvider()?.GetRequiredService<IMBObjectVariableStorage>();
+        private static IMBObjectExtensionDataStore? Instance =>
+            _instance ??= ButterLibSubModule.Instance?.GetServiceProvider()?.GetRequiredService<IMBObjectExtensionDataStore>();
 
         internal static void OnGameEnd() => _instance = null;
 
         /* Variables */
+
+        /**
+         * <summary>
+         * Check whether a variable <paramref name="name"/> is stored upon <paramref name="object"/>.
+         * </summary>
+         * <returns>
+         * <see langword="true"/> if it exists, else <see langword="false"/>.
+         * </returns>
+         * <example>
+         * <code>
+         *  foreach (var skill in SkillObject.All)
+         *      if (!skill.HasVariable("CustomModifiers"))
+         *          RecalculateCustomModifiers(skill);
+         * </code>
+         * </example>
+         * <param name="object">A game object.</param>
+         * <param name="name">The variable's name.</param>
+         */
+        internal static bool HasVariable(this MBObjectBase @object, string name)
+        {
+            if (name is null!)
+                throw new ArgumentNullException(nameof(name));
+            else if (name.Length == 0)
+                throw new ArgumentException("Variable name cannot be empty.", nameof(name));
+
+            return Instance is { } instance && instance.HasVariable(@object, name);
+        }
 
         /**
          * <summary>
@@ -39,12 +67,18 @@ namespace Bannerlord.ButterLib.ObjectSystem.Extensions
          * <typeparam name="T">The type of the variable's value.</typeparam>
          * <param name="object">A game object.</param>
          * <param name="name">The variable's name.</param>
+         * <param name="value">The variable value or <see langword="default"> if nonexistent.</param>
          */
-        public static bool TryGetVariable<T>(this MBObjectBase @object, string name, [MaybeNull] out T value)
+        internal static bool TryGetVariable<T>(this MBObjectBase @object, string name, [MaybeNullWhen(false)][NotNullWhen(true)] out T value)
         {
-            if (Instance is { } instance && instance.TryGetVariable(@object, name, out T val) && val is T typedVal)
+            if (name is null!)
+                throw new ArgumentNullException(nameof(name));
+            else if (name.Length == 0)
+                throw new ArgumentException("Variable name cannot be empty.", nameof(name));
+
+            if (Instance is { } instance && instance.TryGetVariable<T>(@object, name, out var val))
             {
-                value = typedVal;
+                value = val;
                 return true;
             }
 
@@ -52,30 +86,76 @@ namespace Bannerlord.ButterLib.ObjectSystem.Extensions
             return false;
         }
 
-        /// <summary>
-        /// Set the value of the variable <paramref name="name"/> upon <paramref name="object"/>
-        /// to <paramref name="data"/>.
-        /// </summary>
-        /// <example>
-        /// <code>
-        /// myHero.SetVariable("Lovers", new List&lt;Hero&gt; { Hero.MainHero });
-        /// </code>
-        /// </example>
-        /// <typeparam name="T">Type of the variable's value.</typeparam>
-        /// <param name="object">A game object.</param>
-        /// <param name="name">The variable's name.</param>
-        /// <param name="data">The variable's value.</param>
-        public static void SetVariable<T>(this MBObjectBase @object, string name, T data)
+        /** <summary>
+         *  Set the value of the variable <paramref name="name"/> upon <paramref name="object"/>
+         *  to <paramref name="data"/>.
+         *  </summary>
+         *  <example>
+         *  <code>
+         *  // Herein, we're going to find potential lovers (non-spouse consorts) for a lover-less player Hero,
+         *  // and then we'll pick out a few and actually put them into the official Lovers variable of the player
+         *  // with SetVariable.
+         *  //
+         *  // We also assume there's already a HashSet&lt;Hero&gt;-type variable called AttractedTo which is
+         *  // already set on all heroes that are theoretically eligible to marry the player.
+         *
+         *  // Example is for setting a variable outright rather than modifying a retrieved one by reference, so this is only
+         *  // for lover-less players. Modifying existing variables directly by reference is shown at the end.
+         *  if (Hero.MainHero.HasVariable("Lovers"))
+         *      return;
+         *
+         *  // Find possible lovers for the player Hero
+         *  var possibleLovers = Hero.All.Where(h =&gt; h.IsAlive &amp;&amp; // Pretty gross otherwise
+         *                                           Math.Abs(Hero.MainHero.Age - h.Age) &lt; 15f &amp;&amp; // Less than 15 year age gap
+         *                                           // By game rules, whether the character could legally marry the player someday:
+         *                                           Campaign.Current.Models.MarriageModel.IsCoupleSuitableForMarriage(Hero.MainHero, h) &amp;&amp;
+         *                                           // Only those that are attracted to the player (uses another variable, a set of heroes):
+         *                                           h.TryGetVariable("AttractedTo", out HashSet&lt;Hero&gt; attractionSet) &amp;&amp;
+         *                                           attractionSet.Contains(Hero.MainHero) &amp;&amp;
+         *                                           // The attraction must be mutual:
+         *                                           Hero.MainHero.TryGetVariable("AttractedTo", out HashSet&lt;Hero&gt; playerAttractionSet) &amp;&amp;
+         *                                           playerAttractionSet.Contains(h)
+         *                                           );
+         *
+         *  // Take [up to] the 5 youngest of the possible lovers (results in a List&lt;Hero&gt;)
+         *  var lovers = possibleLovers.OrderBy(h =&gt; h.Age).Take(5).ToList();
+         *
+         *  // Set these lovers into the player's variable named "Lovers"
+         *  Hero.MainHero.SetVariable("Lovers", lovers);
+         *
+         *  // Being lovers goes both ways (i.e., if A is B's lover, then B is A's lover):
+         *  foreach (var lover in lovers)
+         *  {
+         *      // Append the player to any potential preexisting lovers that this new lover may have,
+         *      // else set a new Lovers variable upon them which contains only the player.
+         *
+         *      // NOTE: This demonstrates modification of an existing variable by reference, which
+         *      //       often bypasses the need to actually use SetVariable:
+         *      if (lover.TryGetVariable("Lovers", out List&lt;Hero&gt; theirLovers))
+         *          theirLovers.Add(Hero.MainHero); // Modify existing variable by reference
+         *      else
+         *          lover.SetVariable("Lovers", new List&lt;Hero&gt; { Hero.MainHero }); // If var doesn't exist, forced to SetVariable
+         *  }
+         *  </code>
+         *  </example>
+         *  <typeparam name="T">Type of the variable's value.</typeparam>
+         *  <param name="object">A game object.</param>
+         *  <param name="name">The variable's name.</param>
+         *  <param name="data">The variable's value.</param>
+         */
+        internal static void SetVariable<T>(this MBObjectBase @object, string name, T data)
         {
+            if (name is null!)
+                throw new ArgumentNullException(nameof(name));
+            else if (name.Length == 0)
+                throw new ArgumentException("Variable name cannot be empty.", nameof(name));
+
             if (Instance is { } instance)
             {
                 if (data is char @char)
-                {
                     instance.SetVariable(@object, name, @char.ToString());
-                    return;
-                }
-
-                instance.SetVariable(@object, name, data);
+                else
+                    instance.SetVariable(@object, name, data);
             }
         }
 
@@ -89,8 +169,13 @@ namespace Bannerlord.ButterLib.ObjectSystem.Extensions
         /// </example>
         /// <param name="object">A game object.</param>
         /// <param name="name">The variable's name.</param>
-        public static void RemoveVariable(this MBObjectBase @object, string name)
+        internal static void RemoveVariable(this MBObjectBase @object, string name)
         {
+            if (name is null!)
+                throw new ArgumentNullException(nameof(name));
+            else if (name.Length == 0)
+                throw new ArgumentException("Variable name cannot be empty.", nameof(name));
+
             if (Instance is { } instance)
                 instance.RemoveVariable(@object, name);
         }
@@ -109,7 +194,15 @@ namespace Bannerlord.ButterLib.ObjectSystem.Extensions
         /// </example>
         /// <param name="object">A game object.</param>
         /// <param name="name">A string flag.</param>
-        public static bool HasFlag(MBObjectBase @object, string name) => (Instance is { } instance) && instance.HasFlag(@object, name);
+        public static bool HasFlag(this MBObjectBase @object, string name)
+        {
+            if (name is null!)
+                throw new ArgumentNullException(nameof(name));
+            else if (name.Length == 0)
+                throw new ArgumentException("Flag name cannot be empty.", nameof(name));
+
+            return Instance is { } instance && instance.HasFlag(@object, name);
+        }
 
         /// <summary>Set the flag <paramref name="name"/> upon <paramref name="object"/>.</summary>
         /// <example>
@@ -119,8 +212,13 @@ namespace Bannerlord.ButterLib.ObjectSystem.Extensions
         /// </example>
         /// <param name="object">A game object.</param>
         /// <param name="name">A string flag.</param>
-        public static void SetFlag(MBObjectBase @object, string name)
+        public static void SetFlag(this MBObjectBase @object, string name)
         {
+            if (name is null!)
+                throw new ArgumentNullException(nameof(name));
+            else if (name.Length == 0)
+                throw new ArgumentException("Flag name cannot be empty.", nameof(name));
+
             if (Instance is { } instance)
                 instance.SetFlag(@object, name);
         }
@@ -133,8 +231,13 @@ namespace Bannerlord.ButterLib.ObjectSystem.Extensions
         /// </example>
         /// <param name="object">A game object.</param>
         /// <param name="name">A string flag.</param>
-        public static void RemoveFlag(MBObjectBase @object, string name)
+        public static void RemoveFlag(this MBObjectBase @object, string name)
         {
+            if (name is null!)
+                throw new ArgumentNullException(nameof(name));
+            else if (name.Length == 0)
+                throw new ArgumentException("Flag name cannot be empty.", nameof(name));
+
             if (Instance is { } instance)
                 instance.RemoveFlag(@object, name);
         }
