@@ -1,6 +1,5 @@
 ﻿using Bannerlord.ButterLib.Common.Extensions;
 using Bannerlord.ButterLib.Logger.Extensions;
-
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 
@@ -8,10 +7,11 @@ using TaleWorlds.CampaignSystem;
 using TaleWorlds.Core;
 using TaleWorlds.MountAndBlade;
 
-namespace Bannerlord.ButterLib.SaveSystem.Test
+namespace Bannerlord.ButterLib.ObjectSystem.Test
 {
     public sealed class SubModule : MBSubModuleBase
     {
+        private const string Version = "Rev. 5";
         private bool _hasLoaded;
 
         private ILogger _logger = default!;
@@ -31,7 +31,7 @@ namespace Bannerlord.ButterLib.SaveSystem.Test
             {
                 _logger = this.GetServiceProvider().GetRequiredService<ILogger<ButterLibSubModule>>();
 
-                _logger.LogInformationAndDisplay($"Loaded {typeof(SubModule).Namespace}");
+                _logger.LogInformationAndDisplay($"Loaded {typeof(SubModule).Namespace}: {Version}");
                 _hasLoaded = true;
             }
         }
@@ -43,13 +43,9 @@ namespace Bannerlord.ButterLib.SaveSystem.Test
             if (game.GameType is Campaign)
             {
                 CampaignGameStarter initializer = (CampaignGameStarter)starterObject;
-                AddBehaviors(initializer);
+                initializer.AddBehavior(new TestCampaignBehavior());
+                _logger.LogTrace($"Added campaign behavior: {nameof(TestCampaignBehavior)}");
             }
-        }
-
-        private void AddBehaviors(CampaignGameStarter gameInitializer)
-        {
-            gameInitializer.AddBehavior(new TestCampaignBehavior());
         }
     }
 }
